requirements.txt
boto==1.2.3
